+++ resources/js/lib/shareDialog.js
@@ -1,21 +1,16 @@
+import { createToast, ToastType } from './createToast';
+
 export default function shareDialog(shareUrl, shareThumbUrl)
 {
-    function shareClickAnywhereHandler(event)
-    {
-        let shareDialog = document.querySelector(".share-dialog");
-        if(shareDialog == null)
-            return;
-        
-        shareDialog.querySelector("button").dispatchEvent(new Event("click"));
-    }
-
+    let copied = false;
+    
     return new Promise((resolve) =>
     {
-        let shareDialog = document.createElement('div');
+        let dialog = document.createElement('div');
                 
-        shareDialog.classList.toggle("dialog", "true");
-        shareDialog.classList.toggle("share-dialog", "true");
-        shareDialog.innerHTML = `
+        dialog.classList.toggle("dialog", "true");
+        dialog.classList.toggle("share-dialog", "true");
+        dialog.innerHTML = `
             <div class="window">
                 <div class="header">Share Your Masterpiece!</div>
                 <div class="content">
@@ -23,22 +18,33 @@ export default function shareDialog(shareUrl, shareThumbUrl)
                     <div class="input-group">
                         <label>Share URL:</label>
                         <input type="text" id="share-url" value="${shareUrl}" readonly>
-                        <button type="button">Copy</button>
+                        <button class="copy-url" type="button">Copy</button>
                     </div>
+                </div>
+                <div class="footer">
+                    <button class="ok">Close</button>
                 </div>
             </div>`;
         
-        shareDialog.querySelector("button").addEventListener("click", (event) =>
+        dialog.querySelector("button.copy-url").addEventListener("click", (event) =>
         {
             navigator.clipboard.writeText(shareUrl).catch((reason) => console.log(reason));
-            shareDialog.remove();
-            window.removeEventListener("click", shareClickAnywhereHandler);
+            createToast("Copied URL to clipboard.", ToastType.Info);
+            copied = true;
+        });
+
+        dialog.querySelector("button.ok").addEventListener("click", (event) =>
+        {
+            if(!copied)
+            {
+                createToast("Copied URL to clipboard.", ToastType.Info);
+                navigator.clipboard.writeText(shareUrl).catch((reason) => console.log(reason));
+            }
+
+            dialog.remove();
             resolve();
         });
         
-        document.body.appendChild(shareDialog);
-
-        // you're welcome dandistine
-        window.addEventListener("click", shareClickAnywhereHandler);
+        document.body.appendChild(dialog);
     });
 }