+++ resources/js/lib/supportersDialog.js
@@ -1,75 +1,66 @@
 export default function supportersDialog()
 {
-    function supportersClickAnywhereHandler(event)
-    {
-        let supportersDialog = document.querySelector(".dialog.supporters");
-        if(supportersDialog == null)
-            return;
-        
-        if(event.target.tagName == 'A')
-            return;
-        
-        supportersDialog.dispatchEvent(new Event("close-dialog"));
-    }
-
     function renderSupportersDialog(supporters)
     {
-        let supportersDialog = document.createElement("div");
-    
-        supportersDialog.classList.toggle("dialog", true);
-        supportersDialog.classList.toggle("supporters", true);
-
-        let entries = [];
-            
-        if(supporters.length > 0)
+        return new Promise((resolve) =>
         {
-            // sort biggest first
-            supporters.sort((a, b) => b.amount - a.amount);
-            
-            supporters.forEach((entry) =>
+            let dialog = document.createElement("div");
+        
+            dialog.classList.toggle("dialog", true);
+            dialog.classList.toggle("supporters", true);
+
+            let entries = [];
+                
+            if(supporters.length > 0)
             {
-                entries.push(`<div class="name">◀ ${entry.name} ▶</div>`);
-            });
-        }
-        else
-        {
-            entries.push(`<div class="name">◀ No Supporters Yet ▶</div>`);
-        }
+                // sort biggest first
+                supporters.sort((a, b) => b.amount - a.amount);
+                
+                supporters.forEach((entry) =>
+                {
+                    entries.push(`<div class="name">◀ ${entry.name} ▶</div>`);
+                });
+            }
+            else
+            {
+                entries.push(`<div class="name">◀ No Supporters Yet ▶</div>`);
+            }
         
-        supportersDialog.innerHTML = `
-        <div class="window">
-            <div class="header">Patreon Supporters!</div>
-            <div class="content">
-                <h3>PGEtinker would not exist without the support of:</h3>
-                <div class="names">
-                    ${entries.join('')}
+            dialog.innerHTML = `
+            <div class="window">
+                <div class="header">Patreon Supporters!</div>
+                <div class="content">
+                    <h3>PGEtinker would not exist without the support of:</h3>
+                    <div class="names">
+                        ${entries.join('')}
+                    </div>
+                    <a target="_blank" href="https://patreon.com/PGEtinker">
+                        Become a Supporter
+                    </a>                    
                 </div>
-                <a target="_blank" href="https://patreon.com/PGEtinker">
-                    Become a Supporter
-                </a>                    
-            </div>
-        </div>`;
+                <div class="footer">
+                    <button class="ok">Close</button>
+                </div>
+            </div>`;
 
-        supportersDialog.addEventListener("close-dialog", (event) =>
-        {
-            setTimeout(() => window.removeEventListener("click", supportersClickAnywhereHandler), 500);
-            supportersDialog.remove();
+            dialog.querySelector("button.ok").addEventListener("click", (event) =>
+            {
+                dialog.remove();
+                resolve();
+            });
+            
+            document.body.appendChild(dialog);
         });
-        
-        setTimeout(() => window.addEventListener("click", supportersClickAnywhereHandler), 500);
-        document.body.appendChild(supportersDialog);
     }
 
-    return new Promise(() =>
+    return new Promise((resolve) =>
     {
         axios.get("/api/supporters").then((response) =>
         {
-            renderSupportersDialog(response.data.supporters);
+            renderSupportersDialog(response.data.supporters).then(() => resolve());
         }).catch((reason) =>
         {
-            renderSupportersDialog(reason.response.data.supporters);
+            renderSupportersDialog(reason.response.data.supporters).then(() => resolve());
         });
-        
     });
-
 }