+++ routes/web.php
@@ -9,10 +9,6 @@ Route::get('/', function (Request $request)
     return view('home', [
         "code" => "",
         "share_thumb_url" => "",
-
-        // TODO: hook sponsor data into some dynamically loading situation.
-        "navBarSponsorLink" => "https://www.youtube.com/channel/UC-yuWVUplUJZvieEligKBkA",
-        "navBarSponsorText" => "Watch Javidx9, the creator of the olcPixelGameEngine, on Youtube."
     ]);
 });
 
@@ -27,10 +23,6 @@ Route::get('/s/{slug}', function(Request $request, string $slug)
     return view("home", [
         "code" => $code->code,
         "share_thumb_url" => $code->thumb_url,
-        
-        // TODO: hook sponsor data into some dynamically loading situation.
-        "navBarSponsorLink" => "https://www.youtube.com/channel/UC-yuWVUplUJZvieEligKBkA",
-        "navBarSponsorText" => "Watch Javidx9, the creator of the olcPixelGameEngine, on Youtube."
     ]);
 });